D-Bus interface
The D-Bus interface can be used by the applications to directly
register themselves and get the required socket configuration back.

In order to make sure that each application can only register
itself with the settings intended for that application, the
application has to request a D-Bus name like org.detnet.apps.app0
The capability to do this can be controlled for example via
D-Bus policy (currently by means of an explicit user for each
application).

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>