Example with RAW/PACKET support and timestamping
An extended example consisting of a server and a client.
The client sends messages via UDP to the server and
gets a reply. All timestamps on both sides are recorded
and collected on client side.

It supports different socket types for comparison:
- socket(AF_INET, SOCK_DGRAM, 0)
- socket(AF_INET, SOCK_RAW, IPPROTO_UDP)
- socket(AF_INET, SOCK_RAW, IPPROTO_UDP) + IP_HDRINCL
- socket(AF_PACKET, SOCK_DGRAM, htons(ETH_P_IP))
- socket(AF_PACKET, SOCK_RAW, htons(ETH_P_IP))

Also, SCHED_FIFO and CPU affinity can be enabled.

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>