Setup interface via netlink
Use netlink to set the state of both physical and virtual
interface as well as setup the virtual interface including IP address.
This relives from manual configuration and ensures the interfaces
are properly up after registration.
Also, it prevents issues that can occur when the queues
and the eBPF are configured while the interface is up.

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>