fix(configuration): change YANG representation of interface IP address
The IP address of the logical interface was previously
determined from the src-ip-prefix of the app flow.
However, that is only the prefix for flow identification
and in particular it would not be possible to specify
a non-canonical pair of IP address and prefix length
(e.g. 192.168.1.2/24).

Therefore, specify the IP addresses according to RFC 8344.
This also allows to setup more than one IP address for
one logical interface.

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>