feat: dispatch TSN streams using stream identification
Until now, the TSN streams were only identified by the
SO_PRIORITY provided by the application and accordingly
dropped if the provided SO_PRIORITY does not match the
provided SO_TOKEN.

This has three disadvantages:
- Applications can still send to the same TSN stream
  as another application (even though only in a different
  time slot). Depending on the processing at the next
  hop that could still lead to congestion.
- The traffic of applications that do not set the
  SO_PRIORITY will always be handled as best effort.
- If the assignment between SO_PRIORITY and TSN stream
  changes, the application has to be involved to change
  the SO_PRIORITY.
- To set the SO_PRIORITY, the application requires
  at least CAP_NET_RAW priviliges.

Make this more flexible by first identifying the
traffic by destination MAC address and VLAN ID,
then optionally verify the socket token (and drop
otherwise) and finally setting the priority assigned
to this TSN stream by configuration.

The stream identification is comparable to the
Null Stream Identification of IEEE 802.1CB-2017
with the difference that it processes packets
coming down the stack and adds the stream_handle
for sending further down (that gets then processed
by the admission control and stream_handle to
priority mapping), instead of the other way around
as it is the case in network bridges.

Since the eBPF now does not only block certain traffic,
but also dispatches it to different time slots depending
on the stream identification, it got renamed from
guard to dispatcher.

This solves the disadvantages described above:
- The destination MAC address and VLAN ID are also
  protected by the socket token.
- As soon as an application sends to a given TSN
  stream (and it is authorized to do so - that part
  will be improved in an upcoming patch), its
  traffic gets automatically assigned to the correct
  time slot.
- If the SO_PRIORITY to TSN stream assignment changes
  it will be transparently handled by the dispatcher.
- Since the application does not need to set the
  SO_PRIORITY, it does not need special priviliges.

This approach completely ignores the SO_PRIORITY set
by the application which comes with the advantages
above, but it also means that applications that
implement their own more complex timeslot assignment
are no longer supported. If that is a requirement in
the future, we can still either implement more complex
stream identifications in detnetctl (e.g. to include the
mask-and-match scheme) or provide the possibility to assign
a certain subset of priorities to selected TSN streams that
then gets assigned by the application and only verified
by the eBPF.

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>