feat: replace socket tokens with cgroups
Socket tokens are a way to identify a certain socket
in the BPF by a token that was attached to the socket
by the application. By providing these tokens during
the registration procedure, applications are able to
authenticate their own data.

This has the disadvantage that every application needs
to explictly interact with detnetctl to request the
token and attach it to its socket. Therefore, only
applications specifically implemented for the integration
with detnetctl can generate real-time traffic on a
system protected by detnetctl. This excludes in particular
already existing / legacy applications where integrating
detnetctl support might be difficult (e.g. bad
code base or even closed source) or undesirable
(especially applications not originally inteded for the
use in DetNet/TSN contexts).

The other disadvantage is that socket tokens are currently
not supported upstream by Linux and there is no convincing
argument for upstreaming, since there is already a good
alternative: The use of cgroups for traffic identification
in the BPF. Therefore, use that instead of socket tokens.

cgroups are a common approach for managing resources on
Linux systems like CPU and memory and TSN streams are
just a further resource in this sense. So allow only the
application(s) in a cgroup provided during registration
to send traffic to the respective TSN stream.

The disadvantage of using cgroups instead of socket tokens
is that isolating applications in cgroups is a little
more involved than just setting a socket option and the
way how this is done depends on the system (e.g. if
systemd is used or not). But since it is expected that
cgroups are anyway used in the target applications for
resource management (like CPU and memory), this does
not imply additional complexity for the overall system
in contrast to socket tokens.

In addition, in contrast to socket tokens, it is easily
possible to manage cgroups external to the application
itself (without removing the ability to still do it by
the application if wanted). This makes it possible to
integrate applications without explicit detnetctl
support, solving the disadvantage of socket tokens as
outlined above. Therefore, the registration procedure
performed by the example applications (simple and
timestamps) is completely removed with this patch and
detnetctl-run is added that can be called as wrapper
to set up the correct cgroup via systemd and register
the application at detnetctl.

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>