fix: change trafgen interval with send rate
In the script that calls trafgen, the inter-packet timegap (-t)
(not considering the packet transmission time) was mistaken
as send interval (considering the packet transmission time).
Change it to the send rate (-b) and also move the SPDX lines
out of the traffic.cfg into a separate license file, because
it can not be parsed by trafgen otherwise.

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>