docs: clarify dependency of detd on phc2sys
detd uses the CLOCK_TAI to calculate the basetime of the
schedule. This implies that if the CLOCK_TAI and the PHC
have an offset, the basetime might be set in the future
and the NIC will not send packets before that.

While running detd without phc2sys is not really meaningful
anyway for a final application, it might be the case if you
try out several parts invidually as described in the README.

Therefore, clarify the dependency in the README and move
the PTP feature up to setup the system in the right order.

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>