docs(examples): update timestamps example
The timestamps example in its current form no longer
applicable. This in particular includes:

* The way how the application is started has changed.
  Now it is sufficient to just perform a oneshot configuration.
* The timestamps example was in the past tested together with
  a STM32MP1 as receiver so it was much easier to drive the
  server in overload before the client.
* For the same reason, with a modern setup, RSS and XPS can
  lead to very good latencies for the measurement application
  if it results in different queues than the additional high
  traffic.

Therefore, rework the timestamps example to generate much
more traffic and also on all queues (by using random UDP ports)
and adapt the description to the current detnetctl implementation.

Finally, removed the realtime step, because the magnitude
of the effect also highly depends on the used systems and
their configuration.

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>