feat: system setup at startup
Currently, only the PTP configuration takes place at the startup
of detnetctl, while all other configurations (interfaces, queues,
dispatcher) only take place when explicitly requested by the
application.

This is not only inconsistent, but also implies unnecessary
reconfiguration if an application is restarted (including putting
the interface down, leading to all sorts of side effects like
flushing of routing/neighbor tables).

In fact, all information is already available at the start of
detnetctl due to the static configuration, so perform the setup then.

The only thing that needs to be reconfigured later is the explicitly
protection of a TSN stream by providing the cgroup. However, this
even gets optional with the setup at startup and makes the usage
more convenient for use cases where protection is not required.

This patch also prepares the future support of dynamic reconfiguration,
because it allows to decouple the (not yet supported) registration of a
completely new application (that will be routed to the CUC) from applying
a new configuration as a whole to the detnetctl that previously got
calculated centrally at the CNC.

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>