feat: add version to YAML configuration
Significant changes to the structure of the configuration
file might make it impossible to properly parse configuration
files that correspond to an earlier version of the software.

While the decision how to handle such a case need to be made
on a case-by-case basis (still able to parse with new version,
separate parsing for older version or just disallow it),
this always requires to properly identify the version of the
file.

Therefore, add such a field as to the YAML configuration file.

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>