feat: merge YAML with Sysrepo configuration
Currently, the configuration can either exclusively be provided via YAML
or via Sysrepo. This implies that both sets of configuration options are
equivalent. Other configuration options that have no reasonable
representation in YANG can also not be provided via YAML and need to be
provided via CLI parameter (like --protect).

With expected future extensions, this can easily lead to an explosion of
the number and complexity of CLI parameters making it difficult to run
detnetctl.

Therefore, the file-based configuration should no longer be only an
equivalent alternative to Sysrepo, but should also be able to specify
further configuration options in the future that are not represented
by YANG. Those are usually configurations that are specific to the
actual system and are not known externally (such as ones that are
specific to the NIC hardware or to the installed software environment).

Thus, provide the option to provide a configuration file and still
load further configuration from Sysrepo. A configuration option will
always primarily be loaded from Sysrepo and only if it is not available,
fall back to the one loaded from file.

An alternative solution would be to introduce a second configuration
file, but this is likely harder to manage, especially in the case
where only file-based configuration is used and two configuration
trees with overlapping structure need to be maintained and it is
likely often unclear which configuration belongs into which file.

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>