feat: configure cgroup via YAML
The need to configure the cgroups via CLI is
cumbersome, especially when switching between different
configurations for testing purposes.

Therefore, configure the cgroup via YAML file.
The merged configuration allows to also use that
configuration in the YAML file when the remaining
app config is provided by sysrepo.

Note that the default way of providing the cgroup
is still via D-Bus, e.g. via the detnetctl-run tool,
but this is still helpful for testing purposes or
statically engineered setups.

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>