chore: adhere to clippy lints
Rust's clippy is used for static code analysis to improve
code quality and helping to protect against bugs.
This project aims to generate no warnings with strict
clippy setting in the most recent clippy version.
Where this is not possible, the code shall be annotated
accordingly, including an explanation.

After update to clippy 0.1.75, the following lints lead
to errors or warnings in the current code base, so fix
them accordingly.

https://rust-lang.github.io/rust-clippy/master/index.html#/ignored_unit_patterns
https://rust-lang.github.io/rust-clippy/master/index.html#/unnecessary_mut_passed
https://rust-lang.github.io/rust-clippy/master/index.html#/missing_panics_doc

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>