feat: add schedule configuration to YAML
Provide possibility to configure schedule in YAML.
Since the schedule is directly related to an interface,
use a map interface -> schedule.

Also add the configuration entries for TAPRIO that are
not available via YANG since they are more hardware-specific.
This allows to specific the TAPRIO via YAML and the
schedule via sysrepo.

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>