feat: replace detd by direct TAPRIO setup
Currently, the external detd daemon is used to setup the TAPRIO qdisc.
However, it takes a little different approach to configuration than
detnetctl: detd assumes that it is directly interfaced by the
applications which request setting up the system according to their
application requirements. While this is totally fine for TSN end
stations with homogenous traffic-generating applications that are
all able to communicate their requirements to detd, it gets very
complex when the stream requirements come from different heterogenous
sources (such as some from the applications, some from an external
engineering application, some calculated based on a distributed
algorithm...), because it requires to translate them to a format as
if it would have directly been formulated by an application.

This problem is most obvious when detnetctl shall be used to configure
a router that forwards TSN streams and DetNet flows and does not host
any own application. In this case it is usually easier to directly
calculate the time slot schedule externally and push that to the router,
instead of necessarility involving the router in the end-to-end
stream negotiation. Currently, this would imply mimicing an
artificial application that sends packets of a given size and interval
for every stream. In a similar way, this also applies to containerized
end stations that can be modeled as separate internal end-stations
and an additional virtual bridge or router.

Stream negotiation based on the application requirements is of course
still important, but should be performed on a different layer than
the actual configuration of the schedules within the network hardware.
This split aligns well with the different amendments to IEEE 802.1Q:
IEEE 802.1Qdj is meant to be used for the User/Network Interface (UNI)
for communicating the application requirements, while IEEE 802.1Qcw
is meant for pushing (amongst other things) Qbv schedules to bridges
(or with IEEE 802.1DC to other nodes such as end systems or routers).

For the moment, we assume detnetctl is configured with a static schedule
that was calculated externally (e.g by a CNC or even by hand, especially
for demo purposes).

Therefore, remove all uses of detd and replace it with direct
configuration of the TAPRIO qdisc from the schedule provided according
to IEEE 802.1Qcw.

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>