chore: move clippy linter settings to Cargo.toml
Rust's clippy is used for static code analysis to improve
code quality and helping to protect against bugs.
This project aims to generate no warnings with strict
clippy setting in the most recent clippy version.
Where this is not possible, the code shall be annotated
accordingly, including an explanation.

Up to now, the library was checked more strictly than the
binaries, because the respective settings were made in
the src/lib.rs. With the 1.74 release of cargo, the
[lints] section in the Cargo.toml was stabilized, so it is
now possible to move this configuration over to the
Cargo.toml. It now applies to the whole project without
the need to replicate the settings in every binary or
running clippy with additional parameters

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>