chore: do not warn for missing_trait_methods
After moving the linter configuration to the Cargo.toml,
it now also applies to detnetctl-run.rs. Therefore,
https://rust-lang.github.io/rust-clippy/master/index.html#/missing_trait_methods
generates a warning for a missing to_emit_message in
impl dbus::message::SignalArgs for OrgFreedesktopSystemd1ManagerJobRemoved

This is, however, perfectly normal, because default
implementations are already provided by the trait:
https://github.com/diwic/dbus-rs/blob/618262f5e3217cdd173d46d705bbac26c5141e21/dbus/src/message/signalargs.rs#L38

Instead of just allowing it for this special case, remove
the warning completely, because according to the lints documentation
it was anyway only meant for special cases and not to be enabled globally.

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>