feat: minimal GitHub action for clippy
By using GitHub actions, it is easier to make sure that
the static code analysis performed by clippy is always
run in the same environment and in the same way.

Therefore, add a minimal GitHub action that runs clippy.

For now, it does not check the sysrepo feature, because
the upstream packages are still too old, so it would be
necessary to build them from source which is postponed
to a future task.

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>