chore: add more rustc lints
When assembling the list of applicable lints,
only the ones for clippy were considered.
However, rustc itself has several more lints
that are not enabled by default.

Enable all which apply to the current code base,
i.e. do not lead to a warning or error when activated.
Also, add comments for interesting lints that are
currently unstable and should be enabled in the
future when they are stable.

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>