feat: provide priority in configuration
For properly linking the application with the TAPRIO schedule
configuration, the respective priority is needed. Currently,
this is calculated internally of detd, but since it will in
the future directly refer to the priority configured in the
priority to traffic class map, an automatic calculation is
no longer reasonable and it needs to be configurable.

So make it configurable by both YAML and YANG, but do not
already apply it anywhere in this commit.

For YAML, this change is trivial (just adding priority to
the AppConfig), but for YANG this implies significant changes,
because we need to add at least preliminary IEEE 802.1CB support
for actually being able to configure the priority.

This is, however, anyway needed to properly cover my updated
understanding how DetNet IP over TSN shall be configured,
i.e. with an additional stream identification according to
IEEE 802.1CB after the packet has left the DetNet stack.
With that, we can also get rid of the outdated "hack"
(and its explanation) of directly linking the outgoing
virtual interface with the tsn-interface-configuration.

Unfortunately, the tsn-interface-configuration.yang specifying
the stream identification and the time-aware-offset still needs
to be kept and linked by the ingress interface of the app flow.
This is unfortunately a cyclic dependency that we can not clean up
before the direct TAPRIO configuration enables removing the
time-aware-offset for queue setup.

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>