feat: derive convenient access to configuration struct
The change to a merged configuration has introduced some distributed
boilerplate code since it requires to unwrap and check the contents
of each configuration parameter at access, especially since
only the processing module of the respective processing module of the
configuration knows which fields are required and which are optional.
For example, if and only if a TAPRIO qdisc shall be set up, the
parameters regarding the traffic profile are required.

Therefore, implement some macros that generate more convenient
getters for each Option<T>, i.e. one that returns a Result<&T> which
returns an error if the option is None and one that returns a
Option<&T> where the respective configuration parameter is actually
optional.

Also, provide a macro validate_are_some!() that checks if all
provided fields contain a value. This allows to conveniently check
the required values before any action is performed.

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>