feat: replace HashMaps with BTreeMaps in configuration
HashMaps have no order in which the elements are
accessed when iterating over them, while BTreeMaps
are always ordered by their key.

For configuration entries using HashMaps can lead to
a different order in which the entries are processed.
For example, sometimes one interface is set up first
and in the next run it can be another one.
Similarly, the output of the configuration is also
shuffled at each new run.

Therefore, make both the setup and the output more
reproducible by using BTreeMaps instead of HashMaps.

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>