feat: configuration for bridged applications
According to the current DetNet specifications,
all DetNet functionality is only provided by dedicated
bridge/router components, not directly in the ingress/egress
path of an endstation.

detnetctl has ignored this fact in the past and tried
to work around by misusing parts of the YANG model, in
particular by using the ingress_interface of the app flow
in the wrong direction. However, since we also want to
support explicit routers and not only endstations, this
workaround can no longer be reasonably applied.

Therefore, support two types of applications, unbridged
as before and bridged which transmit their traffic via
a veth pair and a bridge/router component.

The TSN dispatcher is only needed for the unbridged
applications because for bridged applications the
mapping of traffic to traffic classes will be performed
by the bridge itself (will be implemented in an upcoming
commit).

Also the cgroup-based guard part of the dispatcher is
not needed because bridged applications shall use
dedicated veths in isolated network namespaces anyway.

Since the DetNet YANG was previously misused for
configuration of unbridged applications, but it is
now needed for configuration of bridged applications
and traffic not originating from this device,
unbridged applications can now only be configured via
YAML, but not sysrepo.

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>