feat: define PCP mapping as interface configuration
Currently, the outgoing PCP is read from the TSN interface configuration.
Before removing that, move the PCP configuration over to the
outgoing interface itself, because it is currently used to set
the PCP for outgoing packets where the priority shall be mapped
to the PCP.

The corresponding YANG configuration can be found in the bridge-port
for the respective interface.

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>