refactor: configure IP addresses in interfaces section
Up to now, setting IP addresses was tightly bundled with the
configuration of the apps an their "logical interface".
To be more flexible for which interface the addresses shall
be configured, move this configuration to the generic
interfaces section. This is also in line with how the
configuration in YANG takes place.

Also, since this enables to setup IP addresses for
interfaces without TAPRIO support, setup the queues
only if the schedule was provided.

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>