feat: multiple interfaces and identifications per stream
To prepare for frame elimination, allow for a single
stream to have traffic incoming via multiple interfaces
and with multiple stream identifications.

In the wording of IEEE 802.1CB, the struct Stream is
a Compound Stream with multiple Member Streams
(represented by the different stream identifications).

It is important to load the same BPF to all the incoming
interfaces of a stream, so they share the maps which
is especially important for the seqrcvy map,
otherwise it would not be possible for one interface
to notice that a packet was already received on the
other interface.

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>