feat: reset FRER history window
FRER requires to reset the history window if no packets were received
for 2 seconds.  In the original xdpfrer implementation this was done
from userspace. However, since we want to keep the data plane running
even if detnetctl itself is no longer running, call the check if a reset
is required into the packet processing itself.

Also, change to bpf_ktime_get_coarse_ns() to reduce the (small) overhead
that comes with this.

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>