feat: disable VLAN offload for XDP
With enabled VLAN offload, the VLAN tag is not present
in the XDP frame data but is needed for stream identification.
Therefore, disable VLAN offload of the relevant interfaces.

Maybe later we could switch to using VLAN xdp hits for this,
but we would likely not gain much by it when we anyway
need the full Ethernet frame for DetNet encapsulation
in the future.

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>