feat: provide interface setup via iproute2 tools
For setting up the interface, the rust-netlink library
is used which directly sends netlink messages to the kernel.
While this was chosen to be a more elegant and direct way
of communicating with the kernel, several disadvantages of
that approach emerged:

* Several required features are not supported by rust-netlink,
  e.g. configuring VLAN offload, configuring TAPRIO qdiscs...

* rust-netlink is not very well maintained, e.g. pull
  requests stay open uncommented for months, so it is
  not possible to add missing features timely. Also,
  this is problematic from a security standpoint.

* For configuring or just displaying interfaces in a different
  network namespace, it is necessary to fork and then unshare
  into that namespace. Since fork without exec only allows
  async-signal-safe functions, this essentially implies the
  actual configuration needs to be done by a different
  executable. So there would be the need to develop a
  dedicated executable to transform command line parameters
  to netlink commands.

This, however, is what is already provided by the iproute2
(and ethtool) tools, which are much better maintained and
more feature-rich. So just take them directly.

This commit adds a functionally equivalent iproute2 version
of the existing netlink module without yet replacing it.

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>