feat: move veth_app to namespace early and validate
With rust-netlink it was not easily possible to perform
configurations on interfaces already in network namespaces.
Therefore, all configuration was made for the host namespace
and then the interface was moved to its namespace.

This also meant an existing setup could not be validated
and that the interface state could not be set up after
the move.

With the use of iproute2 it is now possible to do both
by executing ip within the network namespace. Therefore,
already move the veth_app to the network namespace early
and perform all configurations there.

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>