feat: attach XDP pass in network namespace
A dummy XDP_PASS needs to be attached to the interface
inside the network namespace to enable XDP for the veth
pair. Before, this was trivial since the interface was
moved to the network namespace AFTER installing the XDP
program.

Now, after moving the interface to the namespace early
(or even recycling an existing interface), attaching
directly is no longer possible. Instead, just pin the
program and attach it via the ip command in the
network namespace.

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>