feat: respond with ICMP to too large packets
As stream-based protocol, TCP tries to stuff as many
bytes into a packet as possible. If we just drop these
packets if they are too large, a TCP connection will
never be established.

Therefore, we need to properly respond with a ICMP
message. For this, we just reuse the xdp_ctx of the
packet that was received, fill it with the response
and send it back.

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>