fix: do not wait for setting interface down
When setting interface down, there is no need to
wait until it has fully initialized. Also, while
the speed was also reported for a down link for 6.7 kernel,
it is no longer the case for 6.11. This is expected
behavior, so just skip checking for the speed after
setting link down.

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>